Move embedding-related code to embeddings.ts
